@@ -9,8 +9,7 @@ https://docs.djangoproject.com/en/2.2/ref/settings/
 """
 
 from typing import Dict, List, Tuple, Union
-
-from django.utils.translation import ugettext_lazy as _
+from django.utils.translation import gettext_lazy as _
 
 from server.settings.components import BASE_DIR, config
 
@@ -115,7 +114,6 @@ DATABASES = {
 LANGUAGE_CODE = 'en-us'
 
 USE_I18N = True
-USE_L10N = True
 
 LANGUAGES = (
     ('en', _('English')),
@@ -226,4 +224,8 @@ METADATA_MASK = [
     "difference_text",
     "title_text"
 ]
+
+# The number of experiences to show on the moderation pages
+EXPERIENCES_PER_PAGE = 10
+
 
