@@ -9,6 +9,7 @@ from django.contrib.auth import logout
 from django.shortcuts import redirect, render
 from openhumans.models import OpenHumansMember
 from django.db.models import Q
+from django.urls import reverse
 
 from .models import PublicExperience
 
@@ -28,6 +29,8 @@ from .helpers import (
     delete_PE,
     update_public_experience_db,
     process_trigger_warnings,
+    moderate_page,
+    choose_moderation_redirect,
     extract_triggers_to_show,
     expand_filter,
 )
@@ -285,6 +288,42 @@ def moderate_public_experiences(request):
     else:
         return redirect("main:overview")
 
+def moderation_list(request):
+    """
+    View containing lists of Public Experiences with a given status
+
+    When status is "pending" this includes experiences marked as "unmoderated"
+    or "in review".
+
+    When status is "approved" it includes experiences marked as "approved".
+
+    When status is "rejected" it includes experiences marked as "rejected".
+    """
+    if request.user.is_authenticated and is_moderator(request.user):
+        status = request.GET.get("status", "pending")
+        if status not in ["pending", "approved", "rejected"]:
+            status = "pending"
+
+        if status == "approved":
+            # Search through approved experiences
+            experiences = PublicExperience.objects.filter(
+                moderation_status="approved"
+            ).order_by("created_at")
+        elif status == "rejected":
+            # Search through rejected experiences
+            experiences = PublicExperience.objects.filter(
+                moderation_status="rejected"
+            ).order_by("created_at")
+        else:
+            # Search through unreviewed or in review experiences
+            experiences = PublicExperience.objects.filter(
+                Q(moderation_status="not reviewed") | Q(moderation_status="in review")
+            ).order_by("created_at")
+
+        return moderate_page(request, status, experiences)
+    else:
+        return redirect("main:overview")
+
 # @vcr.use_cassette('tmp/my_stories.yaml', filter_query_parameters=['access_token'])
 def my_stories(request):
     """
@@ -319,6 +358,7 @@ def moderate_experience(request, uuid):
             data = {**unchanged_experience_details, **trigger_details}
 
             moderation_comments = data.pop("moderation_comments", None)
+            moderation_prior = data.pop("moderation_prior", "not moderated")
 
             # get the users OH member id from the model
             user_OH_member = model.open_humans_member
@@ -340,7 +380,8 @@ def moderate_experience(request, uuid):
                 # if user writing E has deauthorized autspaces, delete public experience
                 delete_PE(uuid=uuid, ohmember=user_OH_member)
             # redirect to a new URL:
-            return redirect("main:moderate_public_experiences")
+            status = choose_moderation_redirect(moderation_prior)
+            return redirect("{}?status={}".format(reverse("main:moderation_list"), status))
 
         else:
             experience_title = model.title_text
